Python_19_book_app/requirements.txt
colorama
switchlang
mysql-connector-python
PrettyTable